predict water splashes for the local player
